long untoched changes, hopefully thes work

+++ Lines/HomeWidgetDeleter/app.py
@@ -10,9 +10,7 @@ from values import configs
 logging.basicConfig(level=logging.INFO)
 
 """
-A Lambda function to delete a line from DynamoDB using the lineId passed in the URL.
-It takes an event and context as parameters.
-Returns a dictionary with a status code and a JSON body indicating the result of the deletion.
+
 """
 
 
@@ -20,7 +18,6 @@ def lambda_handler(event, context):
     dynamodb = getDBConnection()
 
     try:
-        # Check if lineId is present in the path parameters
         if "pathParameters" not in event or "id" not in event["pathParameters"]:
             logging.error("Missing widgetId in path parameters")
             return {"statusCode": 400, "body": json.dumps({"message": "Missing lineId in path parameters"})}
@@ -33,13 +30,13 @@ def lambda_handler(event, context):
         # Check if the item exists before attempting to delete
         response = table.get_item(Key={"id": widget_id})
         if "Item" not in response:
-            logging.warning(f"Line with lineId {widget_id} not found")
-            return {"statusCode": 404, "body": json.dumps({"message": "Line not found"})}
-
-        # TODO: The line should belong to the current user first before deleting
+            logging.warning(f"widget with id {widget_id} not found")
+            return {"statusCode": 404, "body": json.dumps({"message": "widget not found"})}
+        if response["Item"]["userId"] != user_id:
+            return {"statusCode": 401, "body": json.dumps({"message": "Unauthorized access, the widget does not belong to the current user"})}
         # Attempt to delete the item
         table.delete_item(Key={"id": widget_id})
-        logging.info(f"Line with lineId {widget_id} successfully deleted")
+        logging.info(f"widget with lineId {widget_id} successfully deleted")
         update_user_updated_at(user_id, int(datetime.utcnow().timestamp() * 1_000))
 
         # Return a success response with the lineId
